refactor: extract chart helpers into shared charts.js
